+++ packages/svelte/src/index.js
@@ -1,5 +1,3 @@
-import Link from "./Link.html";
-import Navigating from "./Navigating.html";
 import curiStore from "./store.js";
 
-export { Link, Navigating, curiStore };
+export { curiStore };