+++ examples/configs/babel/.babelrc.svelte.js
@@ -1,3 +1,4 @@
-module.exoprts = {
-  presets: ["@babel/preset-env"]
+module.exports = {
+  presets: ["@babel/preset-env"],
+  plugins: ["@babel/plugin-syntax-dynamic-import"]
 };