+++ packages/svelte/src/index.js
@@ -1,4 +1,3 @@
-import Link from "./Link.html";
 import curiStore from "./store.js";
 
-export { Link, curiStore };
+export { curiStore };