+++ packages/svelte/tests/link.spec.js
@@ -1,17 +1,20 @@
 import InMemory from "@hickory/in-memory";
-import { curi } from "@curi/router";
+import { curi, prepareRoutes } from "@curi/router";
 import simulant from "simulant";
 import { curiStore } from "@curi/svelte";
 
 import Link from "../src/Link.html";
 
 describe("<Link>", () => {
+  const routes = prepareRoutes([
+    { name: "Home", path: "" },
+    { name: "User", path: "u/:id" },
+    { name: "Not Found", path: "(.*)" }
+  ]);
+
   it("renders an anchor with expected pathname", () => {
     const history = InMemory();
-    const routes = [
-      { name: "Home", path: "" },
-      { name: "Not Found", path: "(.*)" }
-    ];
+
     const router = curi(history, routes);
     const store = curiStore(router);
 
@@ -20,21 +23,17 @@ describe("<Link>", () => {
       target: node,
       store,
       data: {
-        to: "Home"
+        name: "Home"
       }
     });
 
     const a = node.querySelector("a");
     expect(a).not.toBeUndefined();
-    expect(a.pathname).toEqual("/");
+    expect(a.getAttribute("href")).toEqual("/");
   });
 
   it("uses params attribute to generate pathname", () => {
     const history = InMemory();
-    const routes = [
-      { name: "User", path: "u/:id" },
-      { name: "Not Found", path: "(.*)" }
-    ];
     const router = curi(history, routes);
     const store = curiStore(router);
 
@@ -43,21 +42,17 @@ describe("<Link>", () => {
       target: node,
       store,
       data: {
-        to: "User",
+        name: "User",
         params: { id: "1" }
       }
     });
 
     const a = node.querySelector("a");
-    expect(a.pathname).toEqual("/u/1");
+    expect(a.getAttribute("href")).toEqual("/u/1");
   });
 
-  it("falls back to current response's pathname if \"to\" isn't provided", () => {
+  it('uses relative href if "name" isn\'t provided', () => {
     const history = InMemory({ locations: ["/u/2"] });
-    const routes = [
-      { name: "User", path: "u/:id" },
-      { name: "Not Found", path: "(.*)" }
-    ];
     const router = curi(history, routes);
     const store = curiStore(router);
 
@@ -71,16 +66,11 @@ describe("<Link>", () => {
     });
 
     const a = node.querySelector("a");
-    expect(a.pathname).toEqual("/u/2");
-    expect(a.hash).toEqual("#is-a-band");
+    expect(a.getAttribute("href")).toBe("#is-a-band");
   });
 
   it("appends query & hash to end of URI", () => {
     const history = InMemory();
-    const routes = [
-      { name: "Home", path: "" },
-      { name: "Not Found", path: "(.*)" }
-    ];
     const router = curi(history, routes);
     const store = curiStore(router);
 
@@ -89,25 +79,20 @@ describe("<Link>", () => {
       target: node,
       store,
       data: {
-        to: "Home",
+        name: "Home",
         hash: "test",
         query: "one=two"
       }
     });
 
     const a = node.querySelector("a");
-    expect(a.search).toEqual("?one=two");
-    expect(a.hash).toEqual("#test");
+    expect(a.getAttribute("href")).toEqual("/?one=two#test");
   });
 
   describe("clicking a <Link>", () => {
     it("will navigate to the new location", () => {
       const history = InMemory();
       history.navigate = jest.fn();
-      const routes = [
-        { name: "User", path: "u/:id" },
-        { name: "Not Found", path: "(.*)" }
-      ];
       const router = curi(history, routes);
       const store = curiStore(router);
 
@@ -116,7 +101,7 @@ describe("<Link>", () => {
         target: node,
         store,
         data: {
-          to: "User",
+          name: "User",
           params: { id: 1 }
         }
       });
@@ -130,10 +115,6 @@ describe("<Link>", () => {
     it("will ignore modified clicks", () => {
       const history = InMemory();
       history.navigate = jest.fn();
-      const routes = [
-        { name: "User", path: "u/:id" },
-        { name: "Not Found", path: "(.*)" }
-      ];
       const router = curi(history, routes);
       const store = curiStore(router);
 
@@ -142,7 +123,7 @@ describe("<Link>", () => {
         target: node,
         store,
         data: {
-          to: "User",
+          name: "User",
           params: { id: 2 }
         }
       });
@@ -159,10 +140,6 @@ describe("<Link>", () => {
     it("will ignore click if event.defaultPrevented is true", () => {
       const history = InMemory();
       history.navigate = jest.fn();
-      const routes = [
-        { name: "User", path: "u/:id" },
-        { name: "Not Found", path: "(.*)" }
-      ];
       const router = curi(history, routes);
       const store = curiStore(router);
 
@@ -171,7 +148,7 @@ describe("<Link>", () => {
         target: node,
         store,
         data: {
-          to: "User",
+          name: "User",
           params: { id: 3 }
         }
       });
@@ -186,10 +163,6 @@ describe("<Link>", () => {
     it("will ignore click if not done with left mouse button", () => {
       const history = InMemory();
       history.navigate = jest.fn();
-      const routes = [
-        { name: "User", path: "u/:id" },
-        { name: "Not Found", path: "(.*)" }
-      ];
       const router = curi(history, routes);
       const store = curiStore(router);
 
@@ -198,7 +171,7 @@ describe("<Link>", () => {
         target: node,
         store,
         data: {
-          to: "User",
+          name: "User",
           params: { id: 3 }
         }
       });