+++ website/src/pages/Packages/svelte.js
@@ -66,22 +66,24 @@ const store = new Store({...});
 curiStore(router, store);`}
             </CodeBlock>
           </Section>
+
           <Section tag="h3" title={<Cmp>Link</Cmp>} id="link">
             <Explanation>
               <p>
-                In order for the components provided by this package to work,
-                they need to have access to your Curi router object.
+                The <Cmp>Link</Cmp> component is used to create an anchor for
+                navigating to another route.
               </p>
             </Explanation>
             <CodeBlock lang="html">
               {`<div>
-  <Link to='Home'>Home</Link>
-  <Link to='User' params='{ { userID: 5 } }'>
+  <Link name='Home'>Home</Link>
+  <Link name='User' params={{ userID: 5 }}>
     Profile
   </Link>
 </div>
+
 <script>
-  import { Link } from '@curi/svelte';
+  import Link from '@curi/svelte/components/Link.html';
   export default {
     components: { Link }
   }
@@ -89,105 +91,116 @@ curiStore(router, store);`}
             </CodeBlock>
 
             <Section title="Props" id="link-props" tag="h3">
-              <ul>
-                <li>
-                  <Explanation>
-                    <p>
-                      <IJS>to</IJS> - the name of the route to link to
-                    </p>
-                  </Explanation>
-                  <CodeBlock lang="html">
-                    {`<Link to="Home">Home</Link>
+              <Section title="name" id="link-name" tag="h4">
+                <Explanation>
+                  <p>The name of the route to link to.</p>
+                </Explanation>
+                <CodeBlock lang="html">
+                  {`<Link name="Home">Home</Link>
 <!-- <a href="/">Home</a> -->`}
-                  </CodeBlock>
-                </li>
-                <li>
-                  <Explanation>
-                    <p>
-                      <IJS>params</IJS> - any route params for the linked route
-                    </p>
-                  </Explanation>
-                  <CodeBlock lang="html">
-                    {`<Link to="User" params='{ { userID: 5 } }'>
+                </CodeBlock>
+              </Section>
+
+              <Section title="params" id="link-params" tag="h4">
+                <Explanation>
+                  <p>An object of route params for the linked route.</p>
+                </Explanation>
+                <CodeBlock lang="html">
+                  {`<Link name="User" params={{ userID: 5 }}>
   Profile
 </Link>
 <!-- <a href="/user/5">Profile</a> -->`}
-                  </CodeBlock>
-                </li>
-                <li>
-                  <Explanation>
-                    <p>
-                      <IJS>hash</IJS> - the hash for the location to link to
-                    </p>
-                  </Explanation>
-                  <CodeBlock lang="html">
-                    {`<Link to="Home" hash="test">Home</Link>
+                </CodeBlock>
+              </Section>
+
+              <Section title="hash" id="link-hash" tag="h4">
+                <Explanation>
+                  <p>The hash for the location to link to.</p>
+                </Explanation>
+                <CodeBlock lang="html">
+                  {`<Link name="Home" hash="test">Home</Link>
 <!-- <a href="/#test">Home</a> -->`}
-                  </CodeBlock>
-                </li>
-                <li>
+                </CodeBlock>
+              </Section>
+
+              <Section title="query" id="link-query" tag="h4">
+                <Explanation>
+                  <p>The query for the location to link to.</p>
+                </Explanation>
+                <CodeBlock lang="html">
+                  {`<Link name="Home" query="one=1">Home</Link>
+<!-- <a href="/?one=1">Home</a> -->`}
+                </CodeBlock>
+              </Section>
+
+              <Section title="state" id="link-state" tag="h4">
+                <Explanation>
+                  Some (ephemeral) state associated with the location.
+                </Explanation>
+              </Section>
+            </Section>
+
+            <Section tag="h3" title={<Cmp>Navigating</Cmp>} id="navigating">
+              <Explanation>
+                <p>
+                  The <Cmp>Navigating</Cmp> component is used to cancel an
+                  active asynchronous navigation.
+                </p>
+                <p>
+                  A component is passed to <Cmp>Navigating</Cmp>. When there is
+                  an active asynchronous navigation, the component will be given
+                  a <IJS>cancel</IJS> function. When there is not an active
+                  asynchronous navigation, <IJS>cancel</IJS> will be{" "}
+                  <IJS>undefined</IJS>.
+                </p>
+              </Explanation>
+              <CodeBlock lang="html">
+                {`<Navigating component={Cancel} />
+
+<script>
+  import Navigating from "@curi/svelte/components/Navigating.html";
+  import Cancel from "./Cancel";
+
+  export default {
+    components: { Navigating },
+    data() {
+      return { Cancel };
+    }
+  }
+</script>`}
+              </CodeBlock>
+
+              <Section title="Props" id="navigating-props" tag="h3">
+                <Section title="component" id="navigating-component" tag="h4">
                   <Explanation>
                     <p>
-                      <IJS>query</IJS> - the query for the location to link to
+                      A component that receives a <IJS>cancel</IJS> function
+                      when there is an active asynchronous navigation.
                     </p>
                   </Explanation>
                   <CodeBlock lang="html">
-                    {`<Link to="Home" query="one=1">Home</Link>
-<!-- <a href="/?one=1">Home</a> -->`}
+                    {`{#if typeof cancel === "function"}
+  <button on:click="cancelHandler(event, cancel)">
+    Cancel Navigation
+  </button>
+{/if}
+
+<script>
+  export default {
+    methods: {
+      cancelHandler(event, cancel) {
+        event.preventDefault();
+        cancel();
+      }
+    }
+  };
+</script>`}
                   </CodeBlock>
-                </li>
-                <li>
-                  <Explanation>
-                    <IJS>state</IJS> - the state to associated with the location
-                  </Explanation>
-                </li>
-              </ul>
+                </Section>
+              </Section>
             </Section>
           </Section>
         </APIBlock>
-
-        <Section title="Usage" id="usage">
-          <Explanation>
-            <p>
-              For rendering your application, you should should have a base
-              component that has any global layout and uses the response to
-              render the correct component(s). Assuming you setup your store
-              using <IJS>curiStore</IJS>, the components can access the router
-              using <IJS>$router</IJS> and the response and navigation using{" "}
-              <IJS>$curi.response</IJS> and <IJS>$curi.navigation</IJS>.
-            </p>
-          </Explanation>
-          <CodeBlock lang="html">
-            {`<!-- components/App.html -->
-<NavLinks />
-<svelte:component this={$curi.response.body} />
-
-<script>
-  import NavLinks from './NavLinks.html';
-
-  export default {
-	  components: { NavLinks }
-  };
-</script>`}
-          </CodeBlock>
-
-          <Explanation>
-            <p>
-              In your index file, setup an observer to render this component,
-              making sure to pass the store to the component.
-            </p>
-          </Explanation>
-          <CodeBlock>
-            {`import app from './components/app';
-
-const store = curiStore(router);
-
-// use a one time subscriber for the initial render
-router.once(() => {
-  view = new app({ target, store });
-});`}
-          </CodeBlock>
-        </Section>
       </React.Fragment>
     );
   }