+++ packages/svelte/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+* Rename `Link`'s `to` prop to `name`.
+* `Link`s with no name generate a relative `href`.
+
 ## 1.0.0-beta.9
 
 * Add `sideEffects: false` hint for Webpack.