+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Don't bundle components; components should be imported directly from `@curi/svelte/components`.
 * Add `Navigating` component to help cancel asynchronous navigation.
 * Rename `Link`'s `to` prop to `name`.
 * `Link`s with no name generate a relative `href`.